refactor(Views): Renamed/Reorganised model properties for booking
