+++ src/Models/Time/TimeAvailability.cs
@@ -1,8 +1,5 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using form_builder.ViewModels;
-using StockportGovUK.NetStandard.Models.Booking.Response;
 
 namespace form_builder.Models.Time
 {
@@ -17,17 +14,5 @@ namespace form_builder.Models.Time
         public string DateQuestionId { get; set; }
         public ETimePeriod TimePeriodCurrentlySelected { get; set; }
         public ETimePeriod SelectedTimePeriod { get; set; }
-    }
-
-    public class TimePeriod
-    {
-        public ETimePeriod TimeOfDay { get; set; }
-        public List<AppointmentTime> Appointments { get; set; }
-        public bool HasAppointments => Appointments.Any();
-        public DateTime Date { get; set; }
-        public string CurrentValue { get; set; }
-        public string TimeQuestionId { get; set; }
-        public string Id => $"{TimeQuestionId}-{Date.Day}-{TimeOfDay}";
-        public DateTime Value(TimeSpan time) => Date.Add(time);
     }
 }
