+++ src/Models/Time/TimePeriod.cs
@@ -1,0 +1,31 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using form_builder.ViewModels;
+using StockportGovUK.NetStandard.Models.Booking.Response;
+
+namespace form_builder.Models.Time
+{
+    public class TimePeriod
+    {
+        public ETimePeriod TimeOfDay { get; set; }
+        public List<AppointmentTime> Appointments { get; set; }
+        public bool HasAppointments => Appointments.Any();
+        public DateTime Date { get; set; }
+        public string CurrentValue { get; set; }
+        public string TimeQuestionId { get; set; }
+        public string Id => $"{TimeQuestionId}-{Date.Day}-{TimeOfDay}";
+        public DateTime Value(TimeSpan time) => Date.Add(time);
+
+        public string GetClassNameForDivsPosition(int value){
+            switch(value % 3){
+                case 0: 
+                    return "smbc-time__time__label--left";
+                case 1: 
+                    return "smbc-time__time__label--centre";
+                default:
+                    return "smbc-time__time__label--right";
+            }
+        }
+    }
+}
