--- src/ViewModels/TimeAvailabilityViewModel.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using form_builder.Models.Time;
-
-namespace form_builder.ViewModels
-{
-    public class TimeAvailabilityViewModel
-    {
-        public List<TimeAvailability> TimesAvailability { get; set; }
-    }
-}
