+++ src/ViewModels/TimeHeaderViewModel.cs
@@ -4,9 +4,9 @@ namespace form_builder.ViewModels
 {
     public class TimeHeaderViewModel
     {
-        public string AmId { get; set; }
-        public string PmId { get; set; }
+        public string DataAriaControlsIdForMorning { get; set; }
+        public string DataAriaControlsIdForEvening { get; set; }
         public DateTime Date { get; set; }
-        public ETimePeriod Checked { get; set; }
+        public ETimePeriod SelectedTimePeriod { get; set; }
     }
 }
