+++ src/Views/Shared/Booking/Booking.cshtml
@@ -30,10 +30,7 @@
 
     @if (!@Model.IsAppointmentTypeFullDay)
     {
-        var viewModel = new TimeAvailabilityViewModel {
-            TimesAvailability = Model.Times
-        };
-        <partial name="Time" model="@viewModel"/>
+        <partial name="Time" model="@Model.Times"/>
     }
 
     @Html.EndFieldSet()
